pyproject.toml
[project]
name = "music-manager"
version = "0.1.0"
description = "A personal app to help me manage my playlists and other things on Spotify."
readme = "README.md"
authors = [
    { name = "David Ciani", email = "dciani@davidciani.com" }
]

requires-python = ">=3.13"

dependencies = [
    "blend-modes>=2.2.0",
    "click>=8.1.8",
    "numpy>=2.2.3",
    "pillow>=11.1.0",
    "platformdirs",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.0.1",
    "pyyaml>=6.0.2",
    "spotipy>=2.25.1",
]


[project.scripts]
music-manager = "music_manager.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [{include-group = "test"}]
test = [
    "pytest>=8.3.5",
]

[tool.uv.sources]
platformdirs = { path = "../platformdirs" }
